# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ReservedNameRequest:
    properties:
      reserved_name:
        type: string
    required:
    - reserved_name
    type: object
...
  handlers.UserRequest:
    properties:
      dob:
        description: Format "YYYY-MM-DD"
        type: string
      email:
        type: string
      first_name:
        type: string
      middle_name:
        type: string
      password:
        type: string
      surname:
        type: string
    required:
    - email
    type: object
...
  structs.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Kora API server for form management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kora API
  version: "1.0"
paths:
...
  /reserved-name:
    post:
      consumes:
      - application/json
      description: Create a new reserved name
      parameters:
      - description: Reserved Name Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ReservedNameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Create reserved name
      tags:
      - reserved-name
  /reserved-name/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a reserved name by its ID
      parameters:
      - description: Reserved Name ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Delete reserved name
      tags:
      - reserved-name
  /reserved-name/{name}:
    get:
      consumes:
      - application/json
      description: Retrieve reserved names that match or are similar to the provided
        name parameter
      parameters:
      - description: Name to search for
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Get reserved names
      tags:
      - reserved-name
...
  /users:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieve a user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Update user
      tags:
      - users
schemes:
- http
- https
swagger: "2.0"